test/reftest/vendor/blink/fast/dom/DOMException/prototype-object-expected.txt

<p>This tests the prototype chain of DOMException objects.</p><p>On success, you will see a series of "<span class="pass">PASS</span>" messages, followed by "<span class="pass">TEST COMPLETE</span>".</p>
<span><span class="pass">PASS</span> HierarchyRequestError: Node is a host-including inclusive ancestor of parent.</span><br />
<span><span class="pass">PASS</span> Object.prototype.toString.call(e) is "[object DOMException]"</span><br />
<span><span class="pass">PASS</span> Object.prototype.toString.call(e.__proto__) is "[object Object]"</span><br />
<span><span class="pass">PASS</span> Object.prototype.toString.call(e.__proto__.__proto__) is [object Error]</span><br />
<span><span class="pass">PASS</span> function DOMException() { [native code] }</span><br />
<span><span class="pass">PASS</span> e instanceof DOMException is true</span><br />
<span><span class="pass">PASS</span> e instanceof Error is true</span><br />
<span><span class="pass">PASS</span> e.constructor is window.DOMException</span><br />
<span><span class="pass">PASS</span> e.HIERARCHY_REQUEST_ERR is e.constructor.HIERARCHY_REQUEST_ERR</span><br />
<span><span class="pass">PASS</span> e.HIERARCHY_REQUEST_ERR is 3</span><br />
<span><span class="pass">PASS</span> e.code is 3</span><br />
<span><span class="pass">PASS</span> e.name is "HierarchyRequestError"</span><br />
<span><span class="fail">FAIL</span> e.message should be Failed to execute 'appendChild' on 'Node': Nodes of type '#document' may not be inserted inside nodes of type '#document'.. Was Node is a host-including inclusive ancestor of parent..</span><br />
<span><span class="pass">PASS</span> successfullyParsed is true</span><br />
<br /><span class="pass">TEST COMPLETE</span><br />
